requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
blinker==1.4
Click==7.0
colorama==0.4.1
config==0.4.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Caching==1.4.0
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
flask-marshmallow==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
marshmallow==2.17.0
marshmallow-sqlalchemy==0.15.0
more-itertools==4.3.0
pluggy==0.8.0
py==1.7.0
PyMySQL==0.9.3
pytest==4.0.2
python-dotenv==0.10.1
six==1.12.0
SQLAlchemy==1.2.15
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
